refac: move code lenses to their file
