+++ server/src/linter.ts
@@ -1,6 +1,4 @@
 import {
-  CodeLens,
-  CodeLensParams,
   CompletionItem,
   CompletionParams,
   Connection,
@@ -12,20 +10,14 @@ import { Environment, Import, Problem, validate } from 'wollok-ts'
 import { List } from 'wollok-ts/dist/extensions'
 import { completionsForNode } from './functionalities/autocomplete/node-completion'
 import { completeMessages } from './functionalities/autocomplete/send-completion'
-import {
-  getProgramCodeLenses,
-  getTestCodeLenses,
-} from './functionalities/code-lens'
 import { reportValidationMessage } from './functionalities/reporter'
 import { updateDocumentSettings } from './settings'
 import { TimeMeasurer } from './time-measurer'
 import {
-  getWollokFileExtension,
-  packageFromURI,
+  cursorNode,
   trimIn,
 } from './utils/text-documents'
 import { isNodeURI, relativeFilePath, wollokURI } from './utils/vm/wollok'
-import { cursorNode } from './utils/text-documents'
 
 // ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
 // INTERNAL FUNCTIONS
@@ -108,21 +100,6 @@ export const completions = (environment: Environment) => (
   const result = autocompleteMessages ? completeMessages(environment, selectionNode) : completionsForNode(selectionNode)
   timeMeasurer.finalReport()
   return result
-}
-
-export const codeLenses = (environment: Environment) => (params: CodeLensParams): CodeLens[] | null => {
-  const fileExtension = getWollokFileExtension(params.textDocument.uri)
-  const file = packageFromURI(params.textDocument.uri, environment)
-  if (!file) return null
-
-  switch (fileExtension) {
-    case 'wpgm':
-      return getProgramCodeLenses(file)
-    case 'wtest':
-      return getTestCodeLenses(file)
-    default:
-      return null
-  }
 }
 
 export const generateErrorForFile = (connection: Connection, textDocument: TextDocument): void => {
